♻️ [refactor]: 구조 수정, tag 도메인 생성
