--- src/main/java/com/yapp/ios2/fitfty/application/board/BoardFacade.java
@@ -1,4 +1,0 @@
-package com.yapp.ios2.fitfty.application.board;
-
-public class BoardFacade {
-}
