+++ src/main/java/com/yapp/ios2/fitfty/application/tag/TagFacade.java
@@ -1,0 +1,4 @@
+package com.yapp.ios2.fitfty.application.tag;
+
+public class TagFacade {
+}
