+++ src/main/java/com/yapp/ios2/fitfty/domain/picture/Board.java
@@ -1,4 +1,4 @@
-package com.yapp.ios2.fitfty.domain.board;
+package com.yapp.ios2.fitfty.domain.picture;
 
 import com.yapp.ios2.fitfty.domain.AbstractEntity;
