+++ src/main/java/com/yapp/ios2/fitfty/domain/picture/Picture.java
@@ -2,7 +2,7 @@ package com.yapp.ios2.fitfty.domain.picture;
 
 import com.google.common.collect.Lists;
 import com.yapp.ios2.fitfty.domain.AbstractEntity;
-import com.yapp.ios2.fitfty.domain.picture.taggroup.TagGroup;
+import com.yapp.ios2.fitfty.domain.tag.TagGroup;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
